remove `subspace-fraud-proof` crate
