+++ crates/sp-domains-fraud-proof/Cargo.toml
@@ -17,6 +17,7 @@ frame-support = { version = "4.0.0-dev", default-features = false, git = "https:
 hash-db = { version = "0.16.0", default-features = false }
 scale-info = { version = "2.7.0", default-features = false, features = ["derive"] }
 domain-block-preprocessor = { version = "0.1.0", default-features = false, path = "../../domains/client/block-preprocessor", optional = true }
+sc-client-api = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", optional = true }
 sc-executor = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false, optional = true }
 sp-api = { version = "4.0.0-dev", default-features = false, git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
 sp-blockchain = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", optional = true }
@@ -35,6 +36,21 @@ subspace-runtime-primitives = { version = "0.1.0", default-features = false, pat
 trie-db = { version = "0.28.0", default-features = false }
 thiserror = { version = "1.0.48", optional = true }
 
+[dev-dependencies]
+domain-block-builder = { version = "0.1.0", path = "../../domains/client/block-builder" }
+domain-block-preprocessor = { version = "0.1.0", path = "../../domains/client/block-preprocessor" }
+domain-test-service = { version = "0.1.0", path = "../../domains/test/service" }
+futures = "0.3.28"
+pallet-balances = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
+sc-cli = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
+sc-executor = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
+sc-service = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
+subspace-test-client = { version = "0.1.0", path = "../../test/subspace-test-client" }
+subspace-test-service = { version = "0.1.0", path = "../../test/subspace-test-service" }
+subspace-runtime-primitives = { version = "0.1.0", path = "../../crates/subspace-runtime-primitives" }
+tempfile = "3.8.0"
+tokio = "1.32.0"
+
 [features]
 default = ["std"]
 std = [
@@ -44,6 +60,7 @@ std = [
     "hash-db/std",
     "scale-info/std",
     "domain-block-preprocessor",
+    "sc-client-api",
     "sc-executor",
     "sp-api/std",
     "sp-blockchain",
