+++ crates/sp-domains-fraud-proof/src/lib.rs
@@ -17,11 +17,14 @@
 //! Subspace fraud proof primitives for consensus chain.
 #![cfg_attr(not(feature = "std"), no_std)]
 
+#[cfg(feature = "std")]
+pub mod execution_prover;
 pub mod fraud_proof;
 #[cfg(feature = "std")]
 mod host_functions;
 mod runtime_interface;
-pub mod transaction;
+#[cfg(test)]
+mod tests;
 pub mod verification;
 
 use codec::{Decode, Encode};
@@ -34,10 +37,35 @@ pub use runtime_interface::fraud_proof_runtime_interface;
 pub use runtime_interface::fraud_proof_runtime_interface::HostFunctions;
 use sp_api::scale_info::TypeInfo;
 use sp_domains::DomainId;
+use sp_runtime::transaction_validity::{InvalidTransaction, TransactionValidity};
 use sp_runtime_interface::pass_by;
 use sp_runtime_interface::pass_by::PassBy;
 use sp_std::vec::Vec;
 use subspace_core_primitives::Randomness;
+
+/// Custom invalid validity code for the extrinsics in pallet-domains.
+#[repr(u8)]
+pub enum InvalidTransactionCode {
+    BundleEquivocationProof = 101,
+    TransactionProof = 102,
+    ExecutionReceipt = 103,
+    Bundle = 104,
+    FraudProof = 105,
+}
+
+impl From<InvalidTransactionCode> for InvalidTransaction {
+    #[inline]
+    fn from(invalid_code: InvalidTransactionCode) -> Self {
+        InvalidTransaction::Custom(invalid_code as u8)
+    }
+}
+
+impl From<InvalidTransactionCode> for TransactionValidity {
+    #[inline]
+    fn from(invalid_code: InvalidTransactionCode) -> Self {
+        InvalidTransaction::Custom(invalid_code as u8).into()
+    }
+}
 
 /// Request type to fetch required verification information for fraud proof through Host function.
 #[derive(Debug, Decode, Encode, TypeInfo, PartialEq, Eq, Clone)]
