--- crates/sp-domains-fraud-proof/src/transaction.rs
@@ -1,56 +1,0 @@
-use crate::fraud_proof::FraudProof;
-use codec::{Decode, Encode};
-use scale_info::TypeInfo;
-use sp_api::HeaderT;
-use sp_domains::OpaqueBundle;
-use sp_runtime::traits::{Block as BlockT, NumberFor};
-use sp_runtime::transaction_validity::{InvalidTransaction, TransactionValidity};
-use subspace_runtime_primitives::Balance;
-
-/// Custom invalid validity code for the extrinsics in pallet-domains.
-#[repr(u8)]
-pub enum InvalidTransactionCode {
-    BundleEquivocationProof = 101,
-    TransactionProof = 102,
-    ExecutionReceipt = 103,
-    Bundle = 104,
-    FraudProof = 105,
-}
-
-impl From<InvalidTransactionCode> for InvalidTransaction {
-    #[inline]
-    fn from(invalid_code: InvalidTransactionCode) -> Self {
-        InvalidTransaction::Custom(invalid_code as u8)
-    }
-}
-
-impl From<InvalidTransactionCode> for TransactionValidity {
-    #[inline]
-    fn from(invalid_code: InvalidTransactionCode) -> Self {
-        InvalidTransaction::Custom(invalid_code as u8).into()
-    }
-}
-
-/// Object for performing the pre-validation in the transaction pool
-/// before calling into the regular `validate_transaction` runtime api.
-// TODO: Revisit
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug, Decode, Encode, TypeInfo, PartialEq, Eq, Clone)]
-pub enum PreValidationObject<Block, DomainHeader: HeaderT>
-where
-    Block: BlockT,
-{
-    Null,
-    FraudProof(FraudProof<NumberFor<Block>, Block::Hash, DomainHeader>),
-    Bundle(OpaqueBundle<NumberFor<Block>, Block::Hash, DomainHeader, Balance>),
-}
-
-sp_api::decl_runtime_apis! {
-    /// API for extracting the pre-validation objects in the primary chain transaction pool wrapper.
-    pub trait PreValidationObjectApi<DomainHeader: HeaderT> {
-        /// Extract the pre-validation object from the given extrinsic.
-        fn extract_pre_validation_object(
-            extrinsics: Block::Extrinsic,
-        ) -> PreValidationObject<Block, DomainHeader>;
-    }
-}
