--- crates/subspace-fraud-proof/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "subspace-fraud-proof"
-version = "0.1.0"
-authors = ["Liu-Cheng Xu <xuliuchengxlc@gmail.com>"]
-edition = "2021"
-license = "GPL-3.0-or-later"
-homepage = "https://subspace.network"
-repository = "https://github.com/subspace/subspace/"
-description = "Subspace fraud proof utilities"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "3.6.5", features = ["derive"] }
-domain-runtime-primitives = { version = "0.1.0", path = "../../domains/primitives/runtime" }
-domain-block-preprocessor = { version = "0.1.0", path = "../../domains/client/block-preprocessor" }
-futures = "0.3.28"
-hash-db = "0.16.0"
-sc-client-api = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sc-executor = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
-sp-api = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sp-blockchain = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sp-core = { version = "21.0.0", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sp-domains = { version = "0.1.0", path = "../sp-domains" }
-sp-domains-fraud-proof = { version = "0.1.0", path = "../sp-domains-fraud-proof" }
-sp-domain-digests = { version = "0.1.0", path = "../../domains/primitives/digests" }
-sp-runtime = { version = "24.0.0", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sp-state-machine = { version = "0.28.0", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sp-trie = { version = "22.0.0", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-
-[dev-dependencies]
-domain-block-builder = { version = "0.1.0", path = "../../domains/client/block-builder" }
-domain-block-preprocessor = { version = "0.1.0", path = "../../domains/client/block-preprocessor" }
-domain-test-service = { version = "0.1.0", path = "../../domains/test/service" }
-pallet-balances = { version = "4.0.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1" }
-sc-cli = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
-sc-executor = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
-sc-service = { version = "0.10.0-dev", git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e938c6bd2b893d3827d1093cd81baa59a1", default-features = false }
-subspace-test-client = { version = "0.1.0", path = "../../test/subspace-test-client" }
-subspace-test-service = { version = "0.1.0", path = "../../test/subspace-test-service" }
-subspace-runtime-primitives = { version = "0.1.0", path = "../../crates/subspace-runtime-primitives" }
-tempfile = "3.8.0"
-tokio = "1.32.0"
