+++ crates/subspace-service/Cargo.toml
@@ -67,7 +67,6 @@ sp-transaction-pool = { version = "4.0.0-dev", git = "https://github.com/subspac
 static_assertions = "1.1.0"
 subspace-archiving = { version = "0.1.0", path = "../subspace-archiving" }
 subspace-core-primitives = { version = "0.1.0", path = "../subspace-core-primitives" }
-subspace-fraud-proof = { version = "0.1.0", path = "../subspace-fraud-proof" }
 subspace-metrics = { version = "0.1.0", path = "../../shared/subspace-metrics" }
 subspace-networking = { version = "0.1.0", path = "../subspace-networking" }
 subspace-proof-of-space = { version = "0.1.0", path = "../subspace-proof-of-space" }
