+++ crates/subspace-service/src/lib.rs
@@ -28,11 +28,9 @@ pub mod dsn;
 mod metrics;
 pub mod rpc;
 mod sync_from_dsn;
-pub mod tx_pre_validator;
 
 use crate::dsn::{create_dsn_instance, DsnConfigurationError};
 use crate::metrics::NodeMetrics;
-use crate::tx_pre_validator::ConsensusChainTxPreValidator;
 use core::sync::atomic::{AtomicU32, Ordering};
 use cross_domain_message_gossip::cdm_gossip_peers_set_config;
 use domain_runtime_primitives::opaque::{Block as DomainBlock, Header as DomainHeader};
@@ -79,7 +77,6 @@ use sp_consensus_subspace::{
 use sp_core::traits::{CodeExecutor, SpawnEssentialNamed};
 use sp_core::H256;
 use sp_domains::DomainsApi;
-use sp_domains_fraud_proof::transaction::PreValidationObjectApi;
 use sp_domains_fraud_proof::{FraudProofExtension, FraudProofHostFunctionsImpl};
 use sp_externalities::Extensions;
 use sp_objects::ObjectsApi;
@@ -95,15 +92,14 @@ use std::sync::Arc;
 use std::time::Duration;
 use subspace_core_primitives::crypto::kzg::{embedded_kzg_settings, Kzg};
 use subspace_core_primitives::{PotSeed, REWARD_SIGNING_CONTEXT};
-use subspace_fraud_proof::verifier_api::VerifierClient;
 use subspace_metrics::{start_prometheus_metrics_server, RegistryAdapter};
 use subspace_networking::libp2p::multiaddr::Protocol;
 use subspace_networking::libp2p::Multiaddr;
 use subspace_networking::Node;
 use subspace_proof_of_space::Table;
 use subspace_runtime_primitives::opaque::Block;
-use subspace_runtime_primitives::{AccountId, Balance, Hash, Nonce};
-use subspace_transaction_pool::{FullPool, PreValidateTransaction};
+use subspace_runtime_primitives::{AccountId, Balance, Nonce};
+use subspace_transaction_pool::{FullPool, NoopPreValidateTransaction, PreValidateTransaction};
 use tracing::{debug, error, info, Instrument};
 
 // There are multiple places where it is assumed that node is running on 64-bit system, refuse to
@@ -161,21 +157,6 @@ pub type FullClient<RuntimeApi, ExecutorDispatch> =
 
 pub type FullBackend = sc_service::TFullBackend<Block>;
 pub type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
-
-pub type InvalidTransactionProofVerifier<RuntimeApi, ExecutorDispatch> =
-    subspace_fraud_proof::invalid_transaction_proof::InvalidTransactionProofVerifier<
-        Block,
-        FullClient<RuntimeApi, ExecutorDispatch>,
-        Hash,
-        NativeElseWasmExecutor<ExecutorDispatch>,
-        VerifierClient<FullClient<RuntimeApi, ExecutorDispatch>, Block>,
-    >;
-
-pub type FraudProofVerifier<RuntimeApi, ExecutorDispatch> = subspace_fraud_proof::ProofVerifier<
-    Block,
-    InvalidTransactionProofVerifier<RuntimeApi, ExecutorDispatch>,
-    DomainHeader,
->;
 
 /// Subspace networking instantiation variant
 #[derive(Debug)]
@@ -391,15 +372,7 @@ type PartialComponents<RuntimeApi, ExecutorDispatch> = sc_service::PartialCompon
     FullBackend,
     FullSelectChain,
     DefaultImportQueue<Block>,
-    FullPool<
-        Block,
-        FullClient<RuntimeApi, ExecutorDispatch>,
-        ConsensusChainTxPreValidator<
-            Block,
-            FullClient<RuntimeApi, ExecutorDispatch>,
-            FraudProofVerifier<RuntimeApi, ExecutorDispatch>,
-        >,
-    >,
+    FullPool<Block, FullClient<RuntimeApi, ExecutorDispatch>, NoopPreValidateTransaction<Block>>,
     OtherPartialComponents<RuntimeApi, ExecutorDispatch>,
 >;
 
@@ -423,8 +396,7 @@ where
         + TaggedTransactionQueue<Block>
         + SubspaceApi<Block, FarmerPublicKey>
         + DomainsApi<Block, DomainHeader>
-        + ObjectsApi<Block>
-        + PreValidationObjectApi<Block, DomainHeader>,
+        + ObjectsApi<Block>,
     ExecutorDispatch: NativeExecutionDispatch + 'static,
 {
     let telemetry = config
@@ -477,25 +449,11 @@ where
 
     let select_chain = sc_consensus::LongestChain::new(backend.clone());
 
-    let invalid_transaction_proof_verifier = InvalidTransactionProofVerifier::new(
-        client.clone(),
-        executor.clone(),
-        VerifierClient::new(client.clone()),
-    );
-
-    let proof_verifier =
-        subspace_fraud_proof::ProofVerifier::new(Arc::new(invalid_transaction_proof_verifier));
-
-    let tx_pre_validator = ConsensusChainTxPreValidator::new(
-        client.clone(),
-        Box::new(task_manager.spawn_handle()),
-        proof_verifier.clone(),
-    );
     let transaction_pool = subspace_transaction_pool::new_full(
         config,
         &task_manager,
         client.clone(),
-        tx_pre_validator,
+        NoopPreValidateTransaction::<Block>::default(),
     );
 
     let segment_headers_store = SegmentHeadersStore::new(client.clone())
@@ -603,9 +561,7 @@ where
         + HeaderBackend<Block>
         + HeaderMetadata<Block, Error = sp_blockchain::Error>
         + 'static,
-    Client::Api: TaggedTransactionQueue<Block>
-        + DomainsApi<Block, DomainHeader>
-        + PreValidationObjectApi<Block, DomainHeader>,
+    Client::Api: TaggedTransactionQueue<Block> + DomainsApi<Block, DomainHeader>,
     TxPreValidator: PreValidateTransaction<Block = Block> + Send + Sync + Clone + 'static,
 {
     /// Task manager.
@@ -638,14 +594,8 @@ where
     pub transaction_pool: Arc<FullPool<Block, Client, TxPreValidator>>,
 }
 
-type FullNode<RuntimeApi, ExecutorDispatch> = NewFull<
-    FullClient<RuntimeApi, ExecutorDispatch>,
-    ConsensusChainTxPreValidator<
-        Block,
-        FullClient<RuntimeApi, ExecutorDispatch>,
-        FraudProofVerifier<RuntimeApi, ExecutorDispatch>,
-    >,
->;
+type FullNode<RuntimeApi, ExecutorDispatch> =
+    NewFull<FullClient<RuntimeApi, ExecutorDispatch>, NoopPreValidateTransaction<Block>>;
 
 /// Builds a new service for a full client.
 pub async fn new_full<PosTable, RuntimeApi, ExecutorDispatch>(
@@ -670,8 +620,7 @@ where
         + TransactionPaymentApi<Block, Balance>
         + SubspaceApi<Block, FarmerPublicKey>
         + DomainsApi<Block, DomainHeader>
-        + ObjectsApi<Block>
-        + PreValidationObjectApi<Block, DomainHeader>,
+        + ObjectsApi<Block>,
     ExecutorDispatch: NativeExecutionDispatch + 'static,
 {
     let PartialComponents {
