+++ test/subspace-test-client/Cargo.toml
@@ -15,7 +15,7 @@ include = [
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "3.6.5", features = [ "derive" ] }
+codec = { package = "parity-scale-codec", version = "3.6.5", features = ["derive"] }
 evm-domain-test-runtime = { version = "0.1.0", path = "../../domains/test/runtime/evm" }
 fp-evm = { version = "3.0.0-dev", git = "https://github.com/subspace/frontier", rev = "56086daa77802eaa285894bfe4b811be66629c89" }
 futures = "0.3.28"
@@ -34,11 +34,12 @@ sp-runtime = { git = "https://github.com/subspace/polkadot-sdk", rev = "892bf8e9
 subspace-archiving = { path = "../../crates/subspace-archiving" }
 subspace-core-primitives = { path = "../../crates/subspace-core-primitives" }
 subspace-erasure-coding = { path = "../../crates/subspace-erasure-coding" }
-subspace-runtime-primitives = { path = "../../crates/subspace-runtime-primitives" }
 subspace-farmer-components = { path = "../../crates/subspace-farmer-components" }
 subspace-proof-of-space = { path = "../../crates/subspace-proof-of-space" }
+subspace-runtime-primitives = { path = "../../crates/subspace-runtime-primitives" }
 subspace-service = { path = "../../crates/subspace-service" }
 subspace-test-runtime = { version = "0.1.0", path = "../subspace-test-runtime" }
+subspace-transaction-pool = { version = "0.1.0", path = "../../crates/subspace-transaction-pool" }
 zeroize = "1.6.0"
 
 [features]
